+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.33] - 2026-30-04
+### Fixed
+- Fix one-cent payload total drift
+- Fix partial capture error
+### Added
+- Add compatibility for Mealvouchers for Prestashop 9
+
 ## [2.0.32] - 2026-15-04
 ### Fixed
 - Remove CREATED operations from the operations display on the order details page
 ### Changed
 - Correct display of payment method name for split payment transactions
+
 
 ## [2.0.31] - 2026-01-04
 ### Changed